api_tests/requirements.txt
requests==2.31.0
pytest==8.0.0
allure-pytest==2.13.5